Should work as intended now, as the split workGivers was not finished the option for it is behind a #if DEBUG clause
