--- Source/Const.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace HandyUI_PersonalWorkCategories
-{
-    public static class Const
-    {
-        public const string PLACE_FRAME_DEF_NAME = "ConstructFinishFrames"; 
-        public const string PLACE_QUALITY_FRAME_DEF_NAME = "ConstructFinishQualityFrames";
-    }
-}
