+++ Source/HandyUI-PersonalWorkCategories.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31424.327
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HandyUI-PersonalWorkCategories", "HandyUI-PersonalWorkCategories.csproj", "{CBFDFC0A-EEF6-4BCD-A966-7EB8BB69B455}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HandyUI-PersonalWorkCategories", "PersonalWorkCategories\HandyUI-PersonalWorkCategories.csproj", "{CBFDFC0A-EEF6-4BCD-A966-7EB8BB69B455}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
