+++ Source/PersonalWorkCategories/Const.cs
@@ -1,0 +1,7 @@
+﻿namespace HandyUI_PersonalWorkCategories;
+
+public static class Const
+{
+    public const string PLACE_FRAME_DEF_NAME = "ConstructFinishFrames";
+    public const string PLACE_QUALITY_FRAME_DEF_NAME = "ConstructFinishQualityFrames";
+}
