+++ Source/PersonalWorkCategories/Patch/PreClose.cs
@@ -1,0 +1,3 @@
+﻿namespace HandyUI_PersonalWorkCategories.Patch;
+
+internal delegate void PreClose();
